requirements.txt
python-dotenv
pyyaml
sqlalchemy>=2.0
psycopg2-binary
pandas
sqlglot
